+++ src/render.rs
@@ -34,14 +34,14 @@ impl<T: AsRef<str>> Render for crate::Raw<T> {
     }
 }
 
-macro_rules! impl_render_through_display {
+macro_rules! impl_render_through_itoap {
     ($($type:ty)+) => {$(
         impl Render for $type {
             fn render(self, to: &mut String) {
-                self.to_string().render(to)
+                itoap::write_to_string(to, self)
             }
         }
     )+};
 }
 
-impl_render_through_display!(u8 i8 u16 i16 u32 i32 u64 i64 u128 i128 usize isize);
+impl_render_through_itoap!(u8 i8 u16 i16 u32 i32 u64 i64 u128 i128 usize isize);