+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- **Changed**: Render integers through itoap for performance
 
 ## [0.4.0](https://codeberg.org/lsunsi/hiper/src/tag/0.4.0)
 - **Added**: Support let tag for inline declarations